Config file support + change of name
Added config file support and changed the binary name to tost. This
change of name is because tostadora is a long name to write considering
that this is a tool that I constantly use.

Also there is no more signaling to dwmblocks. Since I am not longer
using Dwm (by now), I can not test the new implementation. This will
probably be added in a future update (whenever I return to Dwm).